pyproject.toml
[project]
name = "ipv8-pow-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "cryptography<42",
    "libnacl>=2.1.0",
    "pyipv8>=3.2.0",
]
